requirements.txt
Django>=2.1.0,<2.2.0
djangorestframework>=3.8.0,<3.9.0
mysqlclient>=1.4.0,<1.5.0
flake8>=3.6.0,<3.7.0
Pillow>=5.3.0,<5.4.0
coreapi>=2.3.0,<2.4.0
